Add files via upload
This is the first project of introduction to abstract data types (ADT). A really simple project that consists in calculating the total time that a car has been in a parking lot. To calculate this, we have to create a parking ticket, where in the interface we ask the user to insert his car plate (this has been made in portuguese format, following the XX-33-22 format, where 'XX' can be replaced by any two letters from A to Z and '33' and '22' can be replaced by any two numbers from 0 to 9) following of the time he enters the parking lot and the time he left.
With this info, we can calculate the time the user has been in he parting lot, respecting the formula given (25 minutes has a stay of 1 hour; 1h35m has a stay of 2 hours; 3h01m has a stay of 4 hours and so on).
Giving all of this, we created a typedata struct named 'ticket' and worked our way to the end over there. All the three files 'ticket.c, ticket.h and main.c' can be find in the repository.